+++ src/components/chat/ChatWindow.tsx
@@ -4,17 +4,17 @@ import ChatHistoryList from './ChatHistoryList';
 
 import { X } from 'lucide-react';
 import { ChatWindowProps } from '@/features/chat/types';
-import { useChatManager } from '@/features/chat/hooks/useChatManager';
-import { useChatList } from '@/features/chat/hooks/chatList/useChatList';
+import { useChatManager } from '@/features/chat/hooks/useChatUI';
+
 import { useCreateChat } from '@/features/chat/hooks/chatList/useChatMutation';
-import { useChatMessages, useSendMessage } from '@/features/chat/hooks/chatList/useChatMessage';
+
 import { ChatMessages } from './ChatMessages';
+import { useChatMessages, useSendMessage } from '@/features/chat/hooks/chatMessage/useChatMessage';
+import { useChatList } from '@/features/chat/hooks/chatList/useChatList';
 
 export default function ChatWindow({ isOpen, onClose }: ChatWindowProps) {
-  const { data: chatList = [] } = useChatList();
-
   const { mutate: createChat } = useCreateChat();
-
+  const { data: chatList = [] } = useChatList();
   const generateKoreanTimestamp = () => {
     const now = new Date();
     return `${now.getFullYear()}-${now.getMonth() + 1}-${now.getDate()} ${now.getHours()}:${String(now.getMinutes()).padStart(2, '0')}`;
@@ -22,13 +22,11 @@ export default function ChatWindow({ isOpen, onClose }: ChatWindowProps) {
   const {
     message,
     setMessage,
-    chatHistories,
     currentChatId,
     editingId,
     editTitle,
     setEditTitle,
     loadChat,
-    deleteChat,
     startEditing,
     saveEdit,
     cancelEdit,
@@ -38,7 +36,7 @@ export default function ChatWindow({ isOpen, onClose }: ChatWindowProps) {
     showHistory,
     setShowHistory,
   } = useChatManager();
-  const { data: messages = [], isLoading, refetch } = useChatMessages(currentChatId ?? '');
+  const { data: messages = [], isLoading } = useChatMessages(currentChatId ?? '');
   const { mutate: sendMessage } = useSendMessage();
 
   const handleSubmit = () => {
@@ -48,7 +46,6 @@ export default function ChatWindow({ isOpen, onClose }: ChatWindowProps) {
     setMessage('');
   };
   if (!isOpen) return null;
-  console.log(messages, '1');
   return (
     <div className="fixed bottom-24 right-8 w-[30vw] h-[600px] bg-white rounded-lg shadow-xl border border-gray-200 flex flex-col">
       <div className="flex items-center justify-between p-4 border-b border-gray-200">
@@ -65,10 +62,10 @@ export default function ChatWindow({ isOpen, onClose }: ChatWindowProps) {
             setShowMenu(false);
           }}
           onCreateNewChat={() => {
-            if (chatHistories.length >= 12) {
-              alert('채팅방은 최대 10개까지 생성할 수 있습니다.');
-              return;
-            }
+            // if (chatHistories.length >= 12) {
+            //   alert('채팅방은 최대 10개까지 생성할 수 있습니다.');
+            //   return;
+            // }
             const title = generateKoreanTimestamp();
             createChat(title);
             setShowMenu(false);
@@ -84,7 +81,6 @@ export default function ChatWindow({ isOpen, onClose }: ChatWindowProps) {
             editingId={editingId}
             editTitle={editTitle}
             onLoad={loadChat as any}
-            onDelete={deleteChat}
             onEditStart={startEditing as any}
             onEditChange={setEditTitle}
             onEditSave={saveEdit}
